+++ packages/flutter_chat_view/lib/src/screens/chat_detail_screen.dart
@@ -174,7 +174,6 @@ class _ChatDetailScreenState extends State<ChatDetailScreen> {
                 : "";
 
         return Scaffold(
-          backgroundColor: theme.colorScheme.surface,
           appBar: AppBar(
             backgroundColor: theme.appBarTheme.backgroundColor,
             iconTheme: theme.appBarTheme.iconTheme ??