+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - fix bug where you would route back to the user selection screen insterad of routing back to the chat overview screen
 - Add onPopInvoked callback to the userstory to add custom behaviour for the back button on the chatscreen
 - Handle overflows for users with a long name.
+- Remove the scaffold backgrounds because they should be inherited from the scaffold theme
 
 ## 3.0.0